updated run c code

+++ runcode/runcode.py
@@ -22,27 +22,76 @@ class RunCCode(object):
         self.stdout, self.stderr = a.decode("utf-8"), b.decode("utf-8")
         return result
 
-    def _run_c_prog(self, cmd="./running/a.out",idx=0):
-        #taking custom input
-        my_input = open("./running/input"+str(idx)+".txt","r")
 
-        memory_limit = 5024 #default value, TODO: fetch from the DB
-        time_limit = 2 #default TODO fetch from the DB
+    def fetch_data(self):
+        test_cases = {
+            0:{"input":"2 1 2","output":"NO","points":100},
+            1:{"input":"2 1 3","output":"YES","points":100},
+            2:{"input":"2 7 8","output":"NO","points":100}
+        }
+        return test_cases
+
+
+    def execute_testcase(self,my_input,memory_limit, time_limit,cmd):
         virtualenvcmd = "./timeout -m "+str(memory_limit)
         cmd = cmd+" "+str(time_limit)
-        p = subprocess.Popen(virtualenvcmd+" "+cmd,stdin=my_input,stdout=subprocess.PIPE, stderr=subprocess.PIPE,shell = True)
-        a, b = p.communicate()
-        self.stdout, self.stderr = a.decode("utf-8"), b.decode("utf-8")
-        #checking if memory exceeded
-        arr = self.stderr.split()
-        if(arr[0]=="MEM"):
-            #if the memory limit exceeded
-            ERROR = "MEMORY LIMIT EXCEEDED \n"
-            self.stdout = self.stdout + ERROR
-        elif(arr[0]=="FINISHED"):
-            MSG = "Running successful\n"
-            self.stdout = self.stdout + MSG
+        p = subprocess.Popen(virtualenvcmd+" "+cmd,stdin=subprocess.PIPE,stdout=subprocess.PIPE, stderr=subprocess.PIPE,shell = True)
+        a, b = p.communicate(input = my_input.encode())
+        return a , b
+        
+    def update_test_status(self,test_id,score,time,memory,STATUS,test_case_output):
+        '''
+        fetch the submission for the question
+        update the test status
+        '''
+        test_case_status={"status":STATUS ,"time":time,"memory":memory}
+        test_case_output[test_id] = test_case_status
 
+
+
+
+
+    def _run_c_prog(self, cmd="./running/a.out",idx=0):
+        # taking custom input
+        # this is only if user says cutom input 
+        # my_input = open("./running/input"+str(idx)+".txt","r")
+        memory_limit = 5024 #default value, TODO: fetch from the DB
+        time_limit = 2 #default TODO fetch from the DB
+        '''
+        fetch the inputs from the database
+        loop for the number of test cases
+        compare he output and update the score, time and memory
+        return the score, time , memory taken 
+        to the user in the same format as stored in database
+        '''
+        my_input =  self.fetch_data()
+        score = 0
+        test_case_output = {}
+        for i in range(len(my_input)):
+            a , b = self.execute_testcase(my_input[i]["input"],memory_limit, time_limit,cmd)     
+            self.stdout, self.stderr = a.decode("utf-8"), b.decode("utf-8")
+            target_output=my_input[i]['output']
+            print(self.stdout)
+            if(target_output==self.stdout):
+                score += my_input[i]["points"]
+                print(score)
+            #checking if memory exceeded
+            arr = self.stderr.split()
+            tle_check=self.stdout.split(":")
+            STATUS = "Running Successful"
+            time = "Ok"
+            memory = "Ok"
+            if(tle_check[0]=="TLE"):
+                time = "Not Ok"
+                STATUS = "TLE"+" "+str(time_limit)
+            elif(arr[0]=="MEM"):
+                #if the memory limit exceeded
+                STATUS = "MEMORY LIMIT EXCEEDED \n"
+                memory = "Not Ok"
+            elif(arr[0]=="FINISHED"):
+                STATUS = "Running successful\n"
+            self.update_test_status(i ,score,time,memory,STATUS,test_case_output)
+        print(test_case_output)
     
 
     def run_c_code(self, code=None):
@@ -50,9 +99,14 @@ class RunCCode(object):
             prog_output = "./running/a"+str(self.index)+".out"
             filename = "./running/test"+str(self.index)+".c"
             inputfile = "./running/input"+str(self.index)+".txt"
-            os.remove(prog_output)
-            os.remove(filename)
-            os.remove(inputfile)
+            if os.path.exists(inputfile):
+                os.remove(inputfile)
+            if os.path.exists(filename):
+                os.remove(filename)
+            if os.path.exists(prog_output):
+                os.remove(prog_output)
+            
+            
         idx = self.index
         prog_output = "./running/a"+str(idx)+".out"
         filename = "./running/test"+str(idx)+".c"
@@ -161,3 +215,19 @@ class RunPyCode(object):
             f.write(code)
         self._run_py_prog(filename)
         return self.stderr, self.stdout
+
+
+# #include<stdio.h>
+# int main()
+# {
+# 	int n,a , b;
+# 	scanf("%d",&n);
+# 	scanf("%d",&a);
+# 	scanf("%d",&b);
+# 	int sum=a+b;
+# 	if(sum%2==0)
+# 		printf("Yes");
+# 	else
+# 		printf("No");
+# return 0;
+# }
